extracted com.almasb.fxgl.scene into fxgl-scene module
