+++ fxgl-scene/src/main/java/module-info.java
@@ -1,0 +1,10 @@
+/**
+ * @author Almas Baimagambetov (almaslvl@gmail.com)
+ */
+module com.almasb.fxgl.scene {
+    requires com.almasb.fxgl.core;
+    requires com.almasb.fxgl.input;
+    requires com.almasb.fxgl.time;
+
+    exports com.almasb.fxgl.scene;
+}
