+++ fxgl/pom.xml
@@ -82,6 +82,12 @@
 
         <dependency>
             <groupId>com.github.almasb</groupId>
+            <artifactId>fxgl-scene</artifactId>
+            <version>${fxgl.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>com.github.almasb</groupId>
             <artifactId>fxgl-notification</artifactId>
             <version>${fxgl.version}</version>
         </dependency>
