+++ fxgl/src/main/java/module-info.java
@@ -13,6 +13,7 @@ module com.almasb.fxgl.all {
     requires transitive com.almasb.fxgl.io;
     requires transitive com.almasb.fxgl.media;
     requires transitive com.almasb.fxgl.notification;
+    requires transitive com.almasb.fxgl.scene;
     requires transitive com.almasb.fxgl.time;
     requires transitive com.almasb.fxgl.ui;
 
@@ -32,5 +33,4 @@ module com.almasb.fxgl.all {
     exports com.almasb.fxgl.dsl.components.view;
     exports com.almasb.fxgl.gameplay;
     exports com.almasb.fxgl.saving;
-    exports com.almasb.fxgl.scene;
 }
