+++ fxgl/src/main/kotlin/com/almasb/fxgl/app/CSS.kt
@@ -4,7 +4,7 @@
  * See LICENSE for details.
  */
 
-package com.almasb.fxgl.scene
+package com.almasb.fxgl.app
 
 /**
  * Represents a CSS file.
