+++ fxgl/src/main/kotlin/com/almasb/fxgl/app/DialogSubState.kt
@@ -10,7 +10,6 @@ import com.almasb.fxgl.core.util.Consumer
 import com.almasb.fxgl.core.util.EmptyRunnable
 import com.almasb.fxgl.core.util.Predicate
 import com.almasb.fxgl.dsl.FXGL
-import com.almasb.fxgl.scene.FXGLScene
 import com.almasb.fxgl.scene.SubScene
 import com.almasb.fxgl.ui.DialogBox
 import com.almasb.fxgl.ui.Display
