+++ fxgl/src/main/kotlin/com/almasb/fxgl/app/GameController.kt
@@ -9,14 +9,14 @@ package com.almasb.fxgl.app
 import com.almasb.fxgl.core.collection.PropertyMap
 import com.almasb.fxgl.saving.SaveFile
 import com.almasb.fxgl.saving.UserProfile
-import com.almasb.fxgl.scene.SubScene
+import com.almasb.fxgl.scene.SubSceneStack
 import javafx.beans.property.StringProperty
 
 /**
  *
  * @author Almas Baimagambetov (almaslvl@gmail.com)
  */
-interface GameController : SceneStack {
+interface GameController : SubSceneStack {
 
     fun startNewGame()
 
@@ -49,8 +49,3 @@ interface GameController : SceneStack {
     fun restoreDefaultProfileSettings()
 }
 
-interface SceneStack {
-    fun pushSubScene(subScene: SubScene)
-
-    fun popSubScene()
-}
