+++ fxgl/src/main/kotlin/com/almasb/fxgl/app/IntroScene.kt
@@ -7,7 +7,6 @@
 package com.almasb.fxgl.app
 
 import com.almasb.fxgl.dsl.FXGL
-import com.almasb.fxgl.scene.FXGLScene
 import com.almasb.fxgl.scene.Scene
 
 /**
