+++ fxgl/src/main/kotlin/com/almasb/fxgl/app/LoadingScene.kt
@@ -6,11 +6,8 @@
 
 package com.almasb.fxgl.app
 
-import com.almasb.fxgl.core.util.BiConsumer
-import com.almasb.fxgl.core.util.forEach
 import com.almasb.fxgl.dsl.FXGL
 import com.almasb.fxgl.saving.DataFile
-import com.almasb.fxgl.scene.FXGLScene
 import com.almasb.fxgl.scene.Scene
 import com.almasb.sslogger.Logger
 import javafx.concurrent.Task
