+++ fxgl/src/main/kotlin/com/almasb/fxgl/app/MainWindow.kt
@@ -4,7 +4,6 @@ import com.almasb.fxgl.core.local.Local
 import com.almasb.fxgl.dsl.FXGL
 import com.almasb.sslogger.Logger
 import com.almasb.fxgl.input.MouseEventData
-import com.almasb.fxgl.scene.FXGLScene
 import com.almasb.fxgl.scene.SubScene
 import com.almasb.fxgl.scene.Scene
 import javafx.beans.binding.Bindings
