+++ fxgl/src/main/kotlin/com/almasb/fxgl/app/SceneFactory.kt
@@ -6,9 +6,6 @@
 
 package com.almasb.fxgl.app
 
-import com.almasb.fxgl.scene.FXGLScene
-import com.almasb.fxgl.scene.MenuType
-
 /**
  * Factory for scenes used in FXGL.
  *
