+++ fxgl/src/main/kotlin/com/almasb/fxgl/app/Viewport.kt
@@ -4,7 +4,7 @@
  * See LICENSE for details.
  */
 
-package com.almasb.fxgl.scene
+package com.almasb.fxgl.app
 
 import com.almasb.fxgl.animation.AnimatedValue
 import com.almasb.fxgl.dsl.FXGL
