+++ fxgl/src/main/kotlin/com/almasb/fxgl/dsl/components/KeepOnScreenComponent.kt
@@ -8,7 +8,7 @@ package com.almasb.fxgl.dsl.components
 
 import com.almasb.fxgl.dsl.FXGL
 import com.almasb.fxgl.entity.component.Component
-import com.almasb.fxgl.scene.Viewport
+import com.almasb.fxgl.app.Viewport
 
 /**
  * A component that keeps an entity within the viewport.
