--- fxgl/src/main/kotlin/com/almasb/fxgl/scene/MenuType.kt
@@ -1,16 +1,0 @@
-/*
- * FXGL - JavaFX Game Library. The MIT License (MIT).
- * Copyright (c) AlmasB (almaslvl@gmail.com).
- * See LICENSE for details.
- */
-
-package com.almasb.fxgl.scene
-
-/**
- *
- *
- * @author Almas Baimagambetov (almaslvl@gmail.com)
- */
-enum class MenuType {
-    MAIN_MENU, GAME_MENU
-}
