+++ pom.xml
@@ -21,6 +21,7 @@
         <module>fxgl-entity</module>
         <module>fxgl-animation</module>
         <module>fxgl-ui</module>
+        <module>fxgl-scene</module>
         <module>fxgl</module>
         <module>fxgl-media</module>
         <module>fxgl-input</module>
